+++ backend/bom-bom-server/src/main/java/me/bombom/api/v1/inquiry/controller/InquiryImageController.java
@@ -3,8 +3,8 @@ package me.bombom.api.v1.inquiry.controller;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import me.bombom.api.v1.common.resolver.LoginMember;
-import me.bombom.api.v1.inquiry.dto.InquiryImageUploadResponse;
 import me.bombom.api.v1.inquiry.dto.InquiryRequester;
+import me.bombom.api.v1.inquiry.dto.response.InquiryImageUploadResponse;
 import me.bombom.api.v1.inquiry.resolver.GuestId;
 import me.bombom.api.v1.inquiry.service.InquiryImageService;
 import me.bombom.api.v1.member.domain.Member;
