+++ backend/bom-bom-server/src/main/java/me/bombom/api/v1/inquiry/controller/InquiryMessageController.java
@@ -3,11 +3,11 @@ package me.bombom.api.v1.inquiry.controller;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import me.bombom.api.v1.common.resolver.LoginMember;
-import me.bombom.api.v1.inquiry.dto.InquiryMessagePageResponse;
-import me.bombom.api.v1.inquiry.dto.InquiryMessageResponse;
 import me.bombom.api.v1.inquiry.dto.InquiryRequester;
-import me.bombom.api.v1.inquiry.dto.SendInquiryMessageRequest;
-import me.bombom.api.v1.inquiry.dto.UpdateInquiryMessageRequest;
+import me.bombom.api.v1.inquiry.dto.request.SendInquiryMessageRequest;
+import me.bombom.api.v1.inquiry.dto.request.UpdateInquiryMessageRequest;
+import me.bombom.api.v1.inquiry.dto.response.InquiryMessagePageResponse;
+import me.bombom.api.v1.inquiry.dto.response.InquiryMessageResponse;
 import me.bombom.api.v1.inquiry.resolver.GuestId;
 import me.bombom.api.v1.inquiry.service.InquiryMessageService;
 import me.bombom.api.v1.member.domain.Member;
