+++ backend/bom-bom-server/src/main/java/me/bombom/api/v1/inquiry/controller/InquiryRoomController.java
@@ -3,9 +3,9 @@ package me.bombom.api.v1.inquiry.controller;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import me.bombom.api.v1.common.resolver.LoginMember;
-import me.bombom.api.v1.inquiry.dto.CreateInquiryRoomRequest;
 import me.bombom.api.v1.inquiry.dto.InquiryRequester;
-import me.bombom.api.v1.inquiry.dto.InquiryRoomResponse;
+import me.bombom.api.v1.inquiry.dto.request.CreateInquiryRoomRequest;
+import me.bombom.api.v1.inquiry.dto.response.InquiryRoomResponse;
 import me.bombom.api.v1.inquiry.resolver.GuestId;
 import me.bombom.api.v1.inquiry.service.InquiryRoomService;
 import me.bombom.api.v1.member.domain.Member;
