+++ backend/bom-bom-server/src/main/java/me/bombom/api/v1/inquiry/controller/InquiryRoomControllerApi.java
@@ -5,8 +5,8 @@ import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
-import me.bombom.api.v1.inquiry.dto.CreateInquiryRoomRequest;
-import me.bombom.api.v1.inquiry.dto.InquiryRoomResponse;
+import me.bombom.api.v1.inquiry.dto.request.CreateInquiryRoomRequest;
+import me.bombom.api.v1.inquiry.dto.response.InquiryRoomResponse;
 import me.bombom.api.v1.member.domain.Member;
 import org.springdoc.core.annotations.ParameterObject;
 import org.springframework.data.domain.Page;
