+++ backend/bom-bom-server/src/main/java/me/bombom/api/v1/inquiry/dto/request/UpdateInquiryMessageRequest.java
@@ -1,4 +1,4 @@
-package me.bombom.api.v1.inquiry.dto;
+package me.bombom.api.v1.inquiry.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
