+++ backend/bom-bom-server/src/main/java/me/bombom/api/v1/inquiry/dto/response/InquiryCategoryResponse.java
@@ -1,4 +1,4 @@
-package me.bombom.api.v1.inquiry.dto;
+package me.bombom.api.v1.inquiry.dto.response;
 
 import me.bombom.api.v1.inquiry.domain.InquiryCategory;
