+++ backend/bom-bom-server/src/main/java/me/bombom/api/v1/inquiry/service/InquiryCategoryService.java
@@ -2,7 +2,7 @@ package me.bombom.api.v1.inquiry.service;
 
 import java.util.List;
 import lombok.RequiredArgsConstructor;
-import me.bombom.api.v1.inquiry.dto.InquiryCategoryResponse;
+import me.bombom.api.v1.inquiry.dto.response.InquiryCategoryResponse;
 import me.bombom.api.v1.inquiry.repository.InquiryCategoryRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
