+++ backend/bom-bom-server/src/main/java/me/bombom/api/v1/inquiry/service/InquiryMessageService.java
@@ -10,11 +10,11 @@ import me.bombom.api.v1.common.exception.ErrorDetail;
 import me.bombom.api.v1.inquiry.domain.InquiryMessage;
 import me.bombom.api.v1.inquiry.domain.InquiryMessageImage;
 import me.bombom.api.v1.inquiry.domain.InquiryRoom;
-import me.bombom.api.v1.inquiry.dto.InquiryMessagePageResponse;
-import me.bombom.api.v1.inquiry.dto.InquiryMessageResponse;
 import me.bombom.api.v1.inquiry.dto.InquiryRequester;
-import me.bombom.api.v1.inquiry.dto.SendInquiryMessageRequest;
-import me.bombom.api.v1.inquiry.dto.UpdateInquiryMessageRequest;
+import me.bombom.api.v1.inquiry.dto.request.SendInquiryMessageRequest;
+import me.bombom.api.v1.inquiry.dto.request.UpdateInquiryMessageRequest;
+import me.bombom.api.v1.inquiry.dto.response.InquiryMessagePageResponse;
+import me.bombom.api.v1.inquiry.dto.response.InquiryMessageResponse;
 import me.bombom.api.v1.inquiry.repository.InquiryMessageImageRepository;
 import me.bombom.api.v1.inquiry.repository.InquiryMessageRepository;
 import org.springframework.stereotype.Service;
