+++ backend/bom-bom-server/src/main/java/me/bombom/api/v1/inquiry/service/InquiryRoomService.java
@@ -5,7 +5,7 @@ import me.bombom.api.v1.common.exception.CIllegalArgumentException;
 import me.bombom.api.v1.common.exception.ErrorDetail;
 import me.bombom.api.v1.inquiry.domain.InquiryRoom;
 import me.bombom.api.v1.inquiry.dto.InquiryRequester;
-import me.bombom.api.v1.inquiry.dto.InquiryRoomResponse;
+import me.bombom.api.v1.inquiry.dto.response.InquiryRoomResponse;
 import me.bombom.api.v1.inquiry.repository.InquiryRoomRepository;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
